Refactoring everything to follow clean architecture better
