+++ src/main/java/entity/TranslationFactory.java
@@ -5,8 +5,9 @@ import java.util.List;
  * A factory for creating translations.
  */
 public interface TranslationFactory {
+
     /**
-     * Creates a new translation object
+     * Creates a new signLanguageTranslationDisplay object
      * @param text The text that needs to be translated
      * @param targetLanguage the desired language to be translated into
      * @param translationHistory the previous translated texts
