+++ src/main/java/interface_adapter/speech_to_text/SpeechToTextPresenter.java
@@ -2,20 +2,20 @@ package interface_adapter.speech_to_text;
 
 import use_case.speech_to_text.SpeechToTextOutputBoundary;
 import use_case.speech_to_text.SpeechToTextOutputData;
-import view.SignLanguageView;
+import view.GestureBridgeView;
 
 /**
  * The Presenter that formats and delivers the transcription result to the UI.
  */
 public class SpeechToTextPresenter implements SpeechToTextOutputBoundary {
-    private final SignLanguageView view;
+    private final GestureBridgeView view;
 
     /**
      * Constructor for the SignLanguagePresenter.
      *
      * @param view the view that will display the recognition results
      */
-    public SpeechToTextPresenter(SignLanguageView view) {
+    public SpeechToTextPresenter(GestureBridgeView view) {
         this.view = view;
     }
 
@@ -26,6 +26,6 @@ public class SpeechToTextPresenter implements SpeechToTextOutputBoundary {
      */
     @Override
     public void deliverTranscription(SpeechToTextOutputData outputData) {
-        view.updateTranscriptionText(outputData.getTranscription());
+        view.signLanguageTranscriptionDisplay(outputData.getTranscription());
     }
 }
