+++ src/main/java/use_case/sign_language_translation/SignLanguageTranslationInteractor.java
@@ -3,8 +3,8 @@ package use_case.sign_language_translation;
 import interface_adapter.sign_language_translation.SignLanguageTranslationDataAccessInterface;
 
 /**
- * The interactor responsible for handling the logic of sign language translation.
- * This class processes the translation request and communicates with the data access object and the presenter.
+ * The interactor responsible for handling the logic of sign language signLanguageTranslationDisplay.
+ * This class processes the signLanguageTranslationDisplay request and communicates with the data access object and the presenter.
  */
 public class SignLanguageTranslationInteractor implements SignLanguageTranslationInputBoundary {
 
@@ -24,7 +24,7 @@ public class SignLanguageTranslationInteractor implements SignLanguageTranslatio
     }
 
     /**
-     * Executes the translation logic by taking in user input, translating the text,
+     * Executes the signLanguageTranslationDisplay logic by taking in user input, translating the text,
      * and sending the result to the presenter.
      *
      * @param signLanguageTranslationInputData the input data containing the language and text to be translated
@@ -35,7 +35,7 @@ public class SignLanguageTranslationInteractor implements SignLanguageTranslatio
         final String language = signLanguageTranslationInputData.getLanguage();
         final String text = signLanguageTranslationInputData.getText();
 
-        // Perform the translation using the data access object (DAO)
+        // Perform the signLanguageTranslationDisplay using the data access object (DAO)
         String translatedText = languageDataAccessObject.translate(language, text);
 
         // Prepare the output data with the translated text
