+++ src/main/java/use_case/sign_language_translation/SignLanguageTranslationOutputData.java
@@ -1,8 +1,8 @@
 package use_case.sign_language_translation;
 
 /**
- * A data container class for the result of a sign language translation.
- * It holds the translated text and information about the success or failure of the translation.
+ * A data container class for the result of a sign language signLanguageTranslationDisplay.
+ * It holds the translated text and information about the success or failure of the signLanguageTranslationDisplay.
  */
 public class SignLanguageTranslationOutputData {
 
@@ -10,10 +10,10 @@ public class SignLanguageTranslationOutputData {
     private final boolean useCaseFailed;
 
     /**
-     * Constructor to initialize the output data with translated text and translation status.
+     * Constructor to initialize the output data with translated text and signLanguageTranslationDisplay status.
      *
      * @param translatedText The translated text.
-     * @param useCaseFailed The status of the translation (e.g., success or failure).
+     * @param useCaseFailed The status of the signLanguageTranslationDisplay (e.g., success or failure).
      */
     public SignLanguageTranslationOutputData(String translatedText, boolean useCaseFailed) {
         this.translatedText = translatedText;
@@ -30,9 +30,9 @@ public class SignLanguageTranslationOutputData {
     }
 
     /**
-     * Gets the translation status.
+     * Gets the signLanguageTranslationDisplay status.
      *
-     * @return The status of the translation.
+     * @return The status of the signLanguageTranslationDisplay.
      */
     public boolean useCaseFailed() {
         return useCaseFailed;
